requirements.txt
openai
langchain
PyPDF2
python-dotenv
chromadb
python-telegram-bot
groq
